% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_deviation.R
\name{standard_deviation}
\alias{standard_deviation}
\title{standard_deviation}
\usage{
standard_deviation(x)
}
\arguments{
\item{x}{a list of numbers}
}
\value{
float
}
\description{
standard_deviation
}
\examples{
x <-c(1,2,3,4)
standard_deviation(x)
}

