% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_deviation.R
\name{standarderror}
\alias{standarderror}
\title{standarderror}
\usage{
standarderror(x)
}
\arguments{
\item{x}{a list of numbers}
}
\value{
se a float
}
\description{
standarderror
}
\examples{
x <-c(1,2,3,4) 
standarderror(x)


}

