% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdDev-package.r
\docType{package}
\name{stdDev}
\alias{stdDev}
\alias{stdDev-package}
\title{stdDev: Make standard calculations easy}
\description{
Provides a suite of functions to calculate commonly used values of a list of numbers
}

